+++ CHANGELOG.md
@@ -1,11 +1,13 @@
 # Changelog
 
-## v0.10.8 2026-08-29
+## v0.11.0 2026-08-29
 ### Breaking change
 - Reminder: Please [delete then re-add the integration](https://github.com/flashg1/SolarCharger/wiki/Installation#how-to-delete-and-re-add-the-solarcharger-integration) if upgrading from v0.9.0 or prior.
+- Added charger step current config entity.
 ### Fix
+- Do not call update_ha_state() when restoring values in async_added_to_hass() after reboot.
 ### New feature
-- Configurable charger step current.
+- Support configurable charger step current.
 ### Improvement
 - Added PowerAllocator unit tests.